Moved `reader:{Error,ErrorKind}` to `reader::error::{Error,ErrorKind}` and reexported in `reader`
